refactor(db): relocate database infrastructure to dedicated folder structure
- Move Mongoose schemas from infrastructure/mongoose/schemas to infrastructure/db/mongoose/schemas
- Move repository implementations from infrastructure/repositories to infrastructure/db/repositories
- Move repository tests to infrastructure/db/repositories/tests
- Update all import paths across the application to match the new structure
- Improve organization by grouping all database-related code under the db subfolder
