+++ src/app/api/teams/[teamId]/lineups/route.ts
@@ -1,8 +1,8 @@
 import { NextRequest, NextResponse } from "next/server";
 import { auth } from "@/auth";
-import connectToMongoDB from "@/infrastructure/mongoose/connect-to-mongodb";
-import User from "@/infrastructure/mongoose/schemas/user";
-import Team from "@/infrastructure/mongoose/schemas/team";
+import connectToMongoDB from "@/infrastructure/db/mongoose/connect-to-mongodb";
+import User from "@/infrastructure/db/mongoose/schemas/user";
+import Team from "@/infrastructure/db/mongoose/schemas/team";
 
 export const PATCH = async (
   req: NextRequest,
