+++ src/infrastructure/db/mongoose/schemas/team.ts
