+++ src/infrastructure/db/repositories/index.ts
@@ -1,0 +1,11 @@
+import { BaseMongoRepository } from "@/infrastructure/db/repositories/base.repository.mongo";
+import { UserRepositoryImpl } from "@/infrastructure/db/repositories/user.repository.mongo";
+import { TeamRepositoryImpl } from "@/infrastructure/db/repositories/team.repository.mongo";
+import { RecordRepositoryImpl } from "@/infrastructure/db/repositories/record.repository.mongo";
+
+export {
+  BaseMongoRepository,
+  UserRepositoryImpl,
+  TeamRepositoryImpl,
+  RecordRepositoryImpl,
+};
