--- src/infrastructure/repositories/index.ts
@@ -1,11 +1,0 @@
-import { BaseMongoRepository } from "@/infrastructure/repositories/base.repository.mongo";
-import { UserRepositoryImpl } from "@/infrastructure/repositories/user.repository.mongo";
-import { TeamRepositoryImpl } from "@/infrastructure/repositories/team.repository.mongo";
-import { RecordRepositoryImpl } from "@/infrastructure/repositories/record.repository.mongo";
-
-export {
-  BaseMongoRepository,
-  UserRepositoryImpl,
-  TeamRepositoryImpl,
-  RecordRepositoryImpl,
-};
