+++ src/lib/features/auth/actions/index.ts
@@ -4,7 +4,7 @@ import { hash } from "bcryptjs";
 import { AuthError } from "next-auth";
 import { signIn as authSignIn, signOut } from "@/auth";
 import { DEFAULT_SIGN_IN_REDIRECT } from "@/lib/features/auth/routes";
-import { connectToMongoDB } from "@/infrastructure/mongoose/connect-to-mongodb";
+import { connectToMongoDB } from "@/infrastructure/db/mongoose/connect-to-mongodb";
 import { createVerificationToken } from "@/lib/data/verification-token";
 
 export const signIn = async (provider: string, options?) => {
